requirements.txt
setuptools==40.8.0
wheel==0.32.3
retry==0.9.2
pip==18.1
protobuf==3.9.1
gunicorn==19.9.0
xlrd==1.2.0
google-cloud-datastore==1.10.0
google-cloud-monitoring==0.34.0
APScheduler==3.6.3